chore(api): extract search result serializer past router size cap
The #4542/#4543/#4544 rebase plus title_score emission pushed
routers/search.py over the 2000-counted-line cap. Move
_serialize_search_result to _search_serialize.py with a re-export so
existing import paths keep working; ruff-format daemon.py under the
CI (latest) formatter; regen surface refs.

Refs #4545
